separated ListingDetail logic into custom hook
- created own custom hook that powers the logic of the ListingDetail component
- moved style object in ListingDetail to its own file because it became too large
- minor image rendering updates in Listing component
