+++ swimster-ui/src/components/Listing/style.js
@@ -1,6 +1,6 @@
 export const style = {
     listing: 'flex flex-col text-md font-medium cursor-pointer',
-    images: 'w-full h-64 rounded-xl',
+    images: 'w-full h-64 object-cover rounded-xl',
     listingHeader: '',
     listingTitle: 'text-lg font-semibold mt-1',
     listingAddress: 'text-gray-400',
