+++ swimster-ui/src/components/ListingDetail/style.js
@@ -1,0 +1,29 @@
+export const style = {
+    listingDetail: 'flex flex-col mx-48 mt-20 font-medium text-lg',
+    poolHero: 'flex flex-col',
+    headerContainer: 'flex flex-row justify-between items-center',
+    headerContent: 'flex flex-col',
+    headerContentTitle: 'text-4xl',
+    secondaryHeaderContent: 'flex items-center space-x-4',
+    poolRatings: 'flex',
+    rating: 'pt-0.5',
+    headerContentAddress: 'text-md text-gray-400 pt-0.5',
+    hostProfileImage: 'w-14 h-14 bg-gray-400 rounded-full',
+    imagesContainer: 'grid grid-cols-4 gap-4 mt-5 h-[28rem] items-center',
+    mainImage: 'col-span-2 row-span-2',
+    images: 'bg-gray-300 rounded-xl h-full object-cover',
+    mainPoolDetails: 'flex justify-between items-center mt-4',
+    mainPoolDetailsContent: '',
+    primaryMainPoolDetailsContent: 'text-2xl',
+    secondaryMainPoolDetailsContent: 'flex space-x-4 items-center mt-1',
+    guestsContainer: 'flex space-x-2 items-center',
+    guestsImage: 'w-5 h-5',
+    totalGuests: 'text-sm text-gray-600',
+    poolSocialButtons: 'flex space-x-5',
+    socialButton: 'flex items-center space-x-1 border rounded-xl px-4 py-2 cursor-pointer',
+    socialButtonImage: 'w-8 h-8',
+    socialButtonTitle: '',
+    poolContentContainer: '',
+    morePoolDetails: '',
+    reservationCard: '',
+}
